2022_03
I was too tired for this.
It said that each type was identified by a lower case or upper case letter.
I read that as if a and A was the same, but with different prio.

And when creating the points-map I forgot z, leading to correct test results from the small test but failed on the actual input.

And I got stuck on that both compartments had the same size, which didn't matter.

My original solution was soooo stupid compared to the fixed one.