2024 12
Straight forward part 1, just recursively find more plants of the same
type. For each cell just look at all neighbors to find more cells to
check and to count borders.

Number of sides for part 2 was harder. I quickly realized the number
sides is the same as number of corners (there is a new side started in
every corner) but it then took a while to find the corner-finding
algorithm.