2024 14
My "bug" was not remembering to send pointers to structs, again.

For part 2 I first assumed it would be one of the 100 so I printed all
of them, scrolled and found nothing.

Then I assumed it would take many more steps and was clueless for a
while.
I then tried "the one with the lowest total Manhattan must look
clustered" and that worked, it prints a couple of other ones but finally
(and lastly) the one with the tree.

I renamed ChartAsString to just AsString and created FromString to be
able to store the candidate. Cloning a slice of pointers doesn't realy
work, it will still point to the same elements.