2024 15
Part 2 wasn't that hard.
I added the move-boolean to my pushIt to firt test and then push.
In part 1 I tested and pushed at the same time.

Had a bug where I tried to optimize the pushing but I shouldn't have.

Also added a CopyChart but I'm not using it