2024 16
For the love of ...

Anyway, turns out that append(someSlice, someValue) also changes
someSlice. I knew that. Still failed on it. Again.

Anyway, Part1 was straight forward, BFS with priority queue and a
seen-slice.

Part2 should have been easy and all I did was to store the score found
in part1 and use that to abort searches, but other than that continue
until the queue was empty.