2024 19
Another relaxing day.

Should have tried memoization earlier, went on a rabbit hole to see if
the number of towels could be lowered by dropping some that others could
recreate.

With memoization it was lightning fast.

Part 2 was just about changing the memory from bool to int and to not
exit the recursion on first possible combination but to exhaust all
combinations.

The final solution contains an optimization where a map of towels,
indexed by the first letter, helps minimize the list of towels to test.