2024 20
The initial solution for part1 was to find all walls on the map where
the route passed on both sides. All such finding where "cheats" and once
found I could change the map to block the cheated path, and open the
cheat. Then just run the path-finding again, calculating the save.

For part1 it got obvious that it would be smarter to instead walk the
normal path and from each location find all possible cheats by
continuing the path to the end and seeing if it was allowed to "warp"
there using only the manhattan distance. This automatically eliminates
duplicates as the actual cheat-path isn't in used.

Initially my method to find all cheats returned the cheated path but in
the end it just returns how much was saved on each.